feat: second pass with a wrapper on the `hatchet` instance
